Create DisplayCounter component
