+++ src/components/organisms/index.js
@@ -1,5 +1,7 @@
+import DisplayCounter from "./DisplayCounter";
 import Header from "./Header";
 
 export {
-    Header
+    DisplayCounter,
+    Header,
 };
